+++ apps/admin/backend/src/util/manual_results.ts
